DB: Split HandleNullInterface into two parts
The HandleNullInterface previously encoded two technically separate
properties of an operation: (a) whether the operations needs a wrapper
for NULL values and (b) whether the operation supports safe execution on
invalid (garbage) values. These were only combined as the currently only
user of this information is the EliminateNulls pass.

These properties are now split into separate interfaces. To simplify
code, helpers to declare the method in case of an overriden
implementation are provided (prefixed with "May").