proper lowering of constant db::chars to i32 or util::VarLen32
to support utf-8 chars, we want to represent single chars with 32-bit integers (max len of utf-8 single codepoint) and any larger chars type as a util::VarLen32, aka. string. For this, we now carry information about db::chars right down to the lowering to std where the conversion based on the size happens.