Fix bug under aapcs64 calling convention (Linux ARM)
Under this calling convention, structs consisting of two 64-bit values are passed through the next two free registers.
However, 128-bit integers are only passed in the next aligned pair of registers (e.g., x0+x1 or x2+x3), but not x1+x2.
This difference lead to crashes when calling into the runtime from MLIR where buffers are represented as i128.
For the similar case of VarLen32, the 'fix' has already been applied for similar reasons in the past (on other calling conventions)